requirements.txt
flask
flask_sqlalchemy
flask_login
flask_jwt_extended
authlib
python-dotenv
werkzeug